backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1
httpx>=0.27.0
Pillow>=11.0.0
python-pptx>=1.0.0
numpy>=2.0.0
pydantic>=2.7.1
websockets>=12.0
python-dotenv>=1.0.1
